feat(metering): Optimize deployer metering code by put all metering related code into a single file
