+++ builder/deploy/deployer_test.go
@@ -18,10 +18,8 @@ import (
 	mockbuilder "opencsg.com/csghub-server/_mocks/opencsg.com/csghub-server/builder/deploy/imagebuilder"
 	mockrunner "opencsg.com/csghub-server/_mocks/opencsg.com/csghub-server/builder/deploy/imagerunner"
 	mockScheduler "opencsg.com/csghub-server/_mocks/opencsg.com/csghub-server/builder/deploy/scheduler"
-	mockmq "opencsg.com/csghub-server/_mocks/opencsg.com/csghub-server/builder/mq"
 	mockdb "opencsg.com/csghub-server/_mocks/opencsg.com/csghub-server/builder/store/database"
 	"opencsg.com/csghub-server/builder/deploy/common"
-	"opencsg.com/csghub-server/builder/event"
 	"opencsg.com/csghub-server/builder/loki"
 	"opencsg.com/csghub-server/builder/store/database"
 	"opencsg.com/csghub-server/common/config"
@@ -704,24 +702,6 @@ func TestDeployer_UpdateDeploy(t *testing.T) {
 	require.Nil(t, err)
 }
 
-func TestDeployer_getResourceMap(t *testing.T) {
-	mockSpaceResourceStore := mockdb.NewMockSpaceResourceStore(t)
-	mockSpaceResourceStore.EXPECT().FindAll(mock.Anything).
-		Return([]database.SpaceResource{
-			{ID: 1, Name: "Resource1"},
-			{ID: 2, Name: "Resource2"},
-		}, nil)
-
-	d := &deployer{
-		spaceResourceStore: mockSpaceResourceStore,
-	}
-	resources := d.getResourceMap()
-	require.Equal(t, map[string]string{
-		"1": "Resource1",
-		"2": "Resource2",
-	}, resources)
-}
-
 func TestDeployer_CheckResource(t *testing.T) {
 	config := &config.Config{}
 	config.Runner.VGPUResourceReqKey = "nvidia.com/vgpu"
@@ -814,106 +794,6 @@ func TestDeployer_GetEvaluation(t *testing.T) {
 	r, err := tester.GetEvaluation(ctx, types.EvaluationGetReq{})
 	require.NoError(t, err)
 	require.Equal(t, &types.ArgoWorkFlowRes{ID: 100}, r)
-}
-
-func TestDeployer_getClusterMap(t *testing.T) {
-	clusterStore := mockdb.NewMockClusterInfoStore(t)
-	clusterStore.EXPECT().List(mock.Anything).Return([]database.ClusterInfo{
-		{
-			ClusterID: "cluster1",
-		},
-		{
-			ClusterID: "cluster2",
-		},
-	}, nil)
-
-	d := &deployer{
-		clusterStore: clusterStore,
-	}
-
-	res := d.getClusterMap()
-	require.Equal(t, map[string]database.ClusterInfo{
-		"cluster1": {
-			ClusterID: "cluster1",
-		},
-		"cluster2": {
-			ClusterID: "cluster2",
-		},
-	}, res)
-}
-
-func TestDeployer_CheckHeartbeatTimeout(t *testing.T) {
-	ctx := context.TODO()
-	t.Run("should return true when a cluster times out", func(t *testing.T) {
-		clusterStore := mockdb.NewMockClusterInfoStore(t)
-		cc := &deployer{
-			clusterStore: clusterStore,
-		}
-		clusters := []database.ClusterInfo{
-			{
-				ClusterID: "c1_timed_out",
-				Status:    types.ClusterStatusRunning,
-			},
-			{
-				ClusterID: "c2_active",
-				Status:    types.ClusterStatusRunning,
-			},
-		}
-		clusters[0].UpdatedAt = time.Now().Add(-10 * time.Minute)
-		clusters[1].UpdatedAt = time.Now()
-		clusterStore.EXPECT().ByClusterID(ctx, clusters[0].ClusterID).Once().Return(clusters[0], nil)
-		clusterStore.EXPECT().ByClusterID(ctx, clusters[1].ClusterID).Once().Return(clusters[1], nil)
-		cc.deployConfig.HeartBeatTimeInSec = 5 // seconds
-
-		timedOut, err := cc.CheckHeartbeatTimeout(ctx, "c1_timed_out")
-		require.NoError(t, err)
-		require.True(t, timedOut)
-
-		timedOut, err = cc.CheckHeartbeatTimeout(ctx, "c2_active")
-		require.NoError(t, err)
-		require.False(t, timedOut)
-	})
-
-	t.Run("should return false when no cluster times out", func(t *testing.T) {
-		clusterStore := mockdb.NewMockClusterInfoStore(t)
-		cc := &deployer{
-			clusterStore: clusterStore,
-		}
-		clusters := []database.ClusterInfo{
-			{
-				ClusterID: "c1_active",
-				Status:    types.ClusterStatusRunning,
-			},
-			{
-				ClusterID: "c2_active",
-				Status:    types.ClusterStatusRunning,
-			},
-			{
-				ClusterID: "c3_unavailable",
-				Status:    types.ClusterStatusUnavailable,
-			},
-		}
-
-		clusters[0].UpdatedAt = time.Now().Add(-1 * time.Minute)
-		clusters[1].UpdatedAt = time.Now()
-		clusters[2].UpdatedAt = time.Now().Add(-20 * time.Minute) // Should be timeout
-
-		clusterStore.EXPECT().ByClusterID(ctx, clusters[0].ClusterID).Once().Return(clusters[0], nil)
-		clusterStore.EXPECT().ByClusterID(ctx, clusters[1].ClusterID).Once().Return(clusters[1], nil)
-		clusterStore.EXPECT().ByClusterID(ctx, clusters[2].ClusterID).Once().Return(clusters[2], nil)
-		cc.deployConfig.HeartBeatTimeInSec = 5 * 60 // seconds
-		timedOut, err := cc.CheckHeartbeatTimeout(ctx, "c1_active")
-		require.NoError(t, err)
-		require.False(t, timedOut)
-
-		timedOut, err = cc.CheckHeartbeatTimeout(ctx, "c2_active")
-		require.NoError(t, err)
-		require.False(t, timedOut)
-
-		timedOut, err = cc.CheckHeartbeatTimeout(ctx, "c3_unavailable")
-		require.NoError(t, err)
-		require.True(t, timedOut)
-	})
 }
 
 func TestDeployer_SubmitFinetune(t *testing.T) {
@@ -1014,358 +894,77 @@ func TestDeployer_GetWorkflowLogsNonStream(t *testing.T) {
 	require.NotNil(t, resp)
 }
 
-func TestDeployer_startAcctMeteringRequest(t *testing.T) {
-	now := time.Now()
-	eventTime := now
-
-	t.Run("skip when cluster does not exist", func(t *testing.T) {
-		resMap := map[string]string{"sku1": "resource1"}
-		clusterMap := map[string]database.ClusterInfo{}
-		deploy := database.Deploy{
-			ID:        1,
-			ClusterID: "cluster1",
-			SKU:       "sku1",
-			Type:      types.InferenceType,
+func TestDeployer_CheckHeartbeatTimeout(t *testing.T) {
+	ctx := context.TODO()
+	t.Run("should return true when a cluster times out", func(t *testing.T) {
+		clusterStore := mockdb.NewMockClusterInfoStore(t)
+		cc := &deployer{
+			clusterStore: clusterStore,
 		}
+		clusters := []database.ClusterInfo{
+			{
+				ClusterID: "c1_timed_out",
+				Status:    types.ClusterStatusRunning,
+			},
+			{
+				ClusterID: "c2_active",
+				Status:    types.ClusterStatusRunning,
+			},
+		}
+		clusters[0].UpdatedAt = time.Now().Add(-10 * time.Minute)
+		clusters[1].UpdatedAt = time.Now()
+		clusterStore.EXPECT().ByClusterID(ctx, clusters[0].ClusterID).Once().Return(clusters[0], nil)
+		clusterStore.EXPECT().ByClusterID(ctx, clusters[1].ClusterID).Once().Return(clusters[1], nil)
+		cc.deployConfig.HeartBeatTimeInSec = 5 // seconds
 
-		d := &deployer{}
-		d.startAcctMeteringRequest(resMap, clusterMap, deploy, eventTime)
-		// Should return early without error
+		timedOut, err := cc.CheckHeartbeatTimeout(ctx, "c1_timed_out")
+		require.NoError(t, err)
+		require.True(t, timedOut)
+
+		timedOut, err = cc.CheckHeartbeatTimeout(ctx, "c2_active")
+		require.NoError(t, err)
+		require.False(t, timedOut)
 	})
 
-	t.Run("skip when cluster is not running", func(t *testing.T) {
-		resMap := map[string]string{"sku1": "resource1"}
-		clusterMap := map[string]database.ClusterInfo{
-			"cluster1": {
-				ClusterID: "cluster1",
+	t.Run("should return false when no cluster times out", func(t *testing.T) {
+		clusterStore := mockdb.NewMockClusterInfoStore(t)
+		cc := &deployer{
+			clusterStore: clusterStore,
+		}
+		clusters := []database.ClusterInfo{
+			{
+				ClusterID: "c1_active",
+				Status:    types.ClusterStatusRunning,
+			},
+			{
+				ClusterID: "c2_active",
+				Status:    types.ClusterStatusRunning,
+			},
+			{
+				ClusterID: "c3_unavailable",
 				Status:    types.ClusterStatusUnavailable,
 			},
 		}
-		deploy := database.Deploy{
-			ID:        1,
-			ClusterID: "cluster1",
-			SKU:       "sku1",
-			Type:      types.InferenceType,
-		}
 
-		d := &deployer{}
-		d.startAcctMeteringRequest(resMap, clusterMap, deploy, eventTime)
-		// Should return early without error
-	})
+		clusters[0].UpdatedAt = time.Now().Add(-1 * time.Minute)
+		clusters[1].UpdatedAt = time.Now()
+		clusters[2].UpdatedAt = time.Now().Add(-20 * time.Minute) // Should be timeout
 
-	t.Run("skip when cluster heartbeat timeout", func(t *testing.T) {
-		resMap := map[string]string{"sku1": "resource1"}
-		cluster := database.ClusterInfo{
-			ClusterID: "cluster1",
-			Status:    types.ClusterStatusRunning,
-		}
-		cluster.UpdatedAt = now.Add(-30 * time.Minute) // Very old update
-		clusterMap := map[string]database.ClusterInfo{
-			"cluster1": cluster,
-		}
-		deploy := database.Deploy{
-			ID:        1,
-			ClusterID: "cluster1",
-			SKU:       "sku1",
-			Type:      types.InferenceType,
-		}
+		clusterStore.EXPECT().ByClusterID(ctx, clusters[0].ClusterID).Once().Return(clusters[0], nil)
+		clusterStore.EXPECT().ByClusterID(ctx, clusters[1].ClusterID).Once().Return(clusters[1], nil)
+		clusterStore.EXPECT().ByClusterID(ctx, clusters[2].ClusterID).Once().Return(clusters[2], nil)
+		cc.deployConfig.HeartBeatTimeInSec = 5 * 60 // seconds
+		timedOut, err := cc.CheckHeartbeatTimeout(ctx, "c1_active")
+		require.NoError(t, err)
+		require.False(t, timedOut)
 
-		d := &deployer{
-			deployConfig: common.DeployConfig{
-				HeartBeatTimeInSec: 300, // 5 minutes
-			},
-		}
-		d.startAcctMeteringRequest(resMap, clusterMap, deploy, eventTime)
-		// Should return early without error
-	})
+		timedOut, err = cc.CheckHeartbeatTimeout(ctx, "c2_active")
+		require.NoError(t, err)
+		require.False(t, timedOut)
 
-	t.Run("skip when deploy has no SKU", func(t *testing.T) {
-		resMap := map[string]string{"sku1": "resource1"}
-		cluster := database.ClusterInfo{
-			ClusterID: "cluster1",
-			Status:    types.ClusterStatusRunning,
-		}
-		cluster.UpdatedAt = now
-		clusterMap := map[string]database.ClusterInfo{
-			"cluster1": cluster,
-		}
-		deploy := database.Deploy{
-			ID:        1,
-			ClusterID: "cluster1",
-			SKU:       "", // Empty SKU
-			Type:      types.InferenceType,
-		}
-
-		d := &deployer{}
-		d.startAcctMeteringRequest(resMap, clusterMap, deploy, eventTime)
-		// Should return early without error
-	})
-
-	t.Run("skip when SKU not found in resMap", func(t *testing.T) {
-		resMap := map[string]string{"sku1": "resource1"}
-		cluster := database.ClusterInfo{
-			ClusterID: "cluster1",
-			Status:    types.ClusterStatusRunning,
-		}
-		cluster.UpdatedAt = now
-		clusterMap := map[string]database.ClusterInfo{
-			"cluster1": cluster,
-		}
-		deploy := database.Deploy{
-			ID:        1,
-			ClusterID: "cluster1",
-			SKU:       "unknown_sku",
-			Type:      types.InferenceType,
-		}
-
-		d := &deployer{}
-		d.startAcctMeteringRequest(resMap, clusterMap, deploy, eventTime)
-		// Should return early without error
-	})
-
-	t.Run("skip when invalid deploy type", func(t *testing.T) {
-		resMap := map[string]string{"sku1": "resource1"}
-		cluster := database.ClusterInfo{
-			ClusterID: "cluster1",
-			Status:    types.ClusterStatusRunning,
-		}
-		cluster.UpdatedAt = now
-		clusterMap := map[string]database.ClusterInfo{
-			"cluster1": cluster,
-		}
-		deploy := database.Deploy{
-			ID:        1,
-			ClusterID: "cluster1",
-			SKU:       "sku1",
-			Type:      -1, // Invalid type
-		}
-
-		d := &deployer{}
-		d.startAcctMeteringRequest(resMap, clusterMap, deploy, eventTime)
-		// Should return early without error
-	})
-
-	t.Run("skip when ModelServerless scene", func(t *testing.T) {
-		resMap := map[string]string{"sku1": "resource1"}
-		cluster := database.ClusterInfo{
-			ClusterID: "cluster1",
-			Status:    types.ClusterStatusRunning,
-		}
-		cluster.UpdatedAt = now
-		clusterMap := map[string]database.ClusterInfo{
-			"cluster1": cluster,
-		}
-		hardwareJSON := `{"replicas":1}`
-		deploy := database.Deploy{
-			ID:        1,
-			ClusterID: "cluster1",
-			SKU:       "sku1",
-			Type:      types.ServerlessType,
-			Hardware:  hardwareJSON,
-		}
-
-		d := &deployer{}
-		d.startAcctMeteringRequest(resMap, clusterMap, deploy, eventTime)
-		// Should return early without error
-	})
-
-	t.Run("skip when hardware format is invalid", func(t *testing.T) {
-		resMap := map[string]string{"sku1": "resource1"}
-		cluster := database.ClusterInfo{
-			ClusterID: "cluster1",
-			Status:    types.ClusterStatusRunning,
-		}
-		cluster.UpdatedAt = now
-		clusterMap := map[string]database.ClusterInfo{
-			"cluster1": cluster,
-		}
-		deploy := database.Deploy{
-			ID:        1,
-			ClusterID: "cluster1",
-			SKU:       "sku1",
-			Type:      types.InferenceType,
-			Hardware:  "invalid json",
-		}
-
-		d := &deployer{}
-		d.startAcctMeteringRequest(resMap, clusterMap, deploy, eventTime)
-		// Should return early without error
-	})
-
-	t.Run("success with single node deploy and replica count", func(t *testing.T) {
-		resMap := map[string]string{"sku1": "resource1"}
-		cluster := database.ClusterInfo{
-			ClusterID: "cluster1",
-			Status:    types.ClusterStatusRunning,
-		}
-		cluster.UpdatedAt = now
-		clusterMap := map[string]database.ClusterInfo{
-			"cluster1": cluster,
-		}
-		hardwareJSON := `{"replicas":1}`
-		deploy := database.Deploy{
-			ID:        1,
-			ClusterID: "cluster1",
-			SKU:       "sku1",
-			Type:      types.InferenceType,
-			Hardware:  hardwareJSON,
-			SvcName:   "test-svc",
-			SpaceID:   0,
-			UserUUID:  "user1",
-		}
-
-		mockRunner := mockrunner.NewMockRunner(t)
-		mockRunner.EXPECT().GetReplica(mock.Anything, mock.Anything).
-			Return(&types.ReplicaResponse{
-				ActualReplica:  3,
-				DesiredReplica: 3,
-			}, nil)
-
-		mockMQ := mockmq.NewMockMessageQueue(t)
-		mockMQ.EXPECT().Publish(mock.Anything, mock.Anything).Return(nil)
-
-		d := &deployer{
-			imageRunner: mockRunner,
-			eventPub: &event.EventPublisher{
-				SyncInterval: 5, // 5 minutes
-				MQ:           mockMQ,
-			},
-			deployConfig: common.DeployConfig{
-				HeartBeatTimeInSec: 300,
-				UniqueServiceName:  "test-service",
-			},
-		}
-		d.startAcctMeteringRequest(resMap, clusterMap, deploy, eventTime)
-		// Should complete without error
-	})
-
-	t.Run("success with multi-node deploy (replicas >= 2)", func(t *testing.T) {
-		resMap := map[string]string{"sku1": "resource1"}
-		cluster := database.ClusterInfo{
-			ClusterID: "cluster1",
-			Status:    types.ClusterStatusRunning,
-		}
-		cluster.UpdatedAt = now
-		clusterMap := map[string]database.ClusterInfo{
-			"cluster1": cluster,
-		}
-		hardwareJSON := `{"replicas":2}`
-		deploy := database.Deploy{
-			ID:        1,
-			ClusterID: "cluster1",
-			SKU:       "sku1",
-			Type:      types.InferenceType,
-			Hardware:  hardwareJSON,
-			SvcName:   "test-svc",
-			UserUUID:  "user1",
-		}
-
-		mockMQ := mockmq.NewMockMessageQueue(t)
-		mockMQ.EXPECT().Publish(mock.Anything, mock.Anything).Return(nil)
-
-		d := &deployer{
-			eventPub: &event.EventPublisher{
-				SyncInterval: 5,
-				MQ:           mockMQ,
-			},
-			deployConfig: common.DeployConfig{
-				HeartBeatTimeInSec: 300,
-				UniqueServiceName:  "test-service",
-			},
-		}
-		d.startAcctMeteringRequest(resMap, clusterMap, deploy, eventTime)
-		// Should complete without calling GetReplica
-	})
-
-	t.Run("success with GetReplica returning multiple replicas", func(t *testing.T) {
-		resMap := map[string]string{"sku1": "resource1"}
-		cluster := database.ClusterInfo{
-			ClusterID: "cluster1",
-			Status:    types.ClusterStatusRunning,
-		}
-		cluster.UpdatedAt = now
-		clusterMap := map[string]database.ClusterInfo{
-			"cluster1": cluster,
-		}
-		hardwareJSON := `{"replicas":1}`
-		deploy := database.Deploy{
-			ID:        1,
-			ClusterID: "cluster1",
-			SKU:       "sku1",
-			Type:      types.InferenceType,
-			Hardware:  hardwareJSON,
-			SvcName:   "test-svc",
-			SpaceID:   0,
-			UserUUID:  "user1",
-		}
-
-		mockRunner := mockrunner.NewMockRunner(t)
-		mockRunner.EXPECT().GetReplica(mock.Anything, mock.Anything).
-			Return(&types.ReplicaResponse{
-				ActualReplica:  2,
-				DesiredReplica: 2,
-			}, nil)
-
-		mockMQ := mockmq.NewMockMessageQueue(t)
-		mockMQ.EXPECT().Publish(mock.Anything, mock.Anything).Return(nil)
-
-		d := &deployer{
-			imageRunner: mockRunner,
-			eventPub: &event.EventPublisher{
-				SyncInterval: 5,
-				MQ:           mockMQ,
-			},
-			deployConfig: common.DeployConfig{
-				HeartBeatTimeInSec: 300,
-				UniqueServiceName:  "test-service",
-			},
-		}
-		d.startAcctMeteringRequest(resMap, clusterMap, deploy, eventTime)
-		// Should complete without error
-	})
-
-	t.Run("handle GetReplica error gracefully", func(t *testing.T) {
-		resMap := map[string]string{"sku1": "resource1"}
-		cluster := database.ClusterInfo{
-			ClusterID: "cluster1",
-			Status:    types.ClusterStatusRunning,
-		}
-		cluster.UpdatedAt = now
-		clusterMap := map[string]database.ClusterInfo{
-			"cluster1": cluster,
-		}
-		hardwareJSON := `{"replicas":1}`
-		deploy := database.Deploy{
-			ID:        1,
-			ClusterID: "cluster1",
-			SKU:       "sku1",
-			Type:      types.InferenceType,
-			Hardware:  hardwareJSON,
-			SvcName:   "test-svc",
-			SpaceID:   0,
-			UserUUID:  "user1",
-		}
-
-		mockRunner := mockrunner.NewMockRunner(t)
-		mockRunner.EXPECT().GetReplica(mock.Anything, mock.Anything).
-			Return(nil, errors.New("replica error"))
-
-		mockMQ := mockmq.NewMockMessageQueue(t)
-		mockMQ.EXPECT().Publish(mock.Anything, mock.Anything).Return(nil)
-
-		d := &deployer{
-			imageRunner: mockRunner,
-			eventPub: &event.EventPublisher{
-				SyncInterval: 5,
-				MQ:           mockMQ,
-			},
-			deployConfig: common.DeployConfig{
-				HeartBeatTimeInSec: 300,
-				UniqueServiceName:  "test-service",
-			},
-		}
-		d.startAcctMeteringRequest(resMap, clusterMap, deploy, eventTime)
-		// Should use default replicaCount of 1
+		timedOut, err = cc.CheckHeartbeatTimeout(ctx, "c3_unavailable")
+		require.NoError(t, err)
+		require.True(t, timedOut)
 	})
 }
